FIXES weights for no-weights case, dataset.distribution cumulative plots
In the case, where no weights are given to the dataset, all
category._weights should be None. However, this has shown to create
problems with several functions (like category.integrated_rate) which
expect array data. It is fixed by setting making the categeory.weights
property creating weights with np.ones on the fly in the no-weight case.
Another bug fixed is the handling of the cumulative plot in the
VariableDistributionPlot called by dataset.distribution. It if fixed by
creating a default axes_locator in the case where keywords cumulative
and/or ratio are given.