REFACTOR - pass through dataset plotting dictionary to categories
When a dataset is loaded, a dictionary providing style information
for visualization can be set for that dataset.
However, categories of this dataset did not care about the settings
in this dataset-wide dictionary.
This behaviour has been changed. If a category of a dataset is plotted,
it will look up in the dataset dictionary if there is any style
information available.